Always return unicode with hexdigest() (#6313)
### Problem
`sha1.hexdigest()` returns bytes in Py2, but unicode in Py3. 

This inconsistency increases the complexity of trying to support Py2 and Py3. We should always return the same type regardless of interpreter.

### Solution
Adopt the Python 3 semantics, i.e. always return unicode. We are trying to write Python 3 first code, so this is the preferred solution. It also requires less boilerplate with things like string interpolation.

If we really want to be returning bytes, the [docs](https://docs.python.org/3/library/hashlib.html#hashlib.hash.hexdigest) explain we should then be using `digest()` instead of `hexdigest()`, as hexdigest is designed to get a unicode string based off of `digest`.

--

This does not modify local variables when it does not matter.