Switch to Py2 and Py3 shards. (#6289)
### Problem

In order to avoid backsliding as changes are made in service of #6062, we should begin running a travis shard that uses python 3 even before it has gone completely green.

### Solution

Repurpose one of the unit test shards to run with python 3, with abbreviated logging and `|| exit 0`. This should allow us to track the total failure count under python 3, and attempt to push it down.