Process execution: Create symlink to JDK on demand
This creates a symlink `.jdk` pointing at the passed JDK, if one was
passed.

This is a hack to allow for a remote execution platform to provide a
pointer at its pre-installed JDK without needing to pollute absolute
paths into the action definition.

Long-term, we want the JDK to be part of the input directory for the
action, but that's a lot of work, and we want to experiment with
pre-installed JDKs in the mean time.