Requirements on language-specific sources should be optional. (#6375)
E.g., a repo may have java without scala or vice versa.

This only works today because we implicitly expect codegen tasks
(which produce java and scala), and the deferred sources mapper task, to
be registered.  But not having them registered is a totally
reasonable thing to do (e.g., it improves start up time because
there are fewer options to parse).

This extends the logic of https://github.com/pantsbuild/pants/pull/6357.

The goal is to be able to run pants with just the backends you want.
In fact, once this is possible we should remove all the default registrations
(keeping, say, just graph_info and project_info) and require you to
opt in to whatever lang-specific backends you need.

Debatably, many more, possibly even all, requirements should be optional.
But that's a bigger issue, for another day.