Fix --binaries-path-by-id fingerprinting error (#6413)
Fixes #6394. Any attempt to override the global bootstrap option `--binaries-path-by-id` currently fails with a `TypeError`, as described in that ticket.

- Override the `default` method on our JSONEncoder subclass to handle non-list iterables.
- Encode each key of the `fingerprint()` method's `option_val` argument with `stable_json_sha1()`, if `option_val` is a dict (this produces string keys).

The `--binaries-path-by-id` option can now be overridden or extended from the environment or `pants.ini`. Passing this argument on the command line still fails because the option is parsed as "the `--path-by-id` option in the `binaries` scope".