Cancel running work when entering the fork context (#6464)
As described in #6426: when an error occurs prefork with `pantsd` and a request fails fast, work that is still running in the engine can race the shutdown of all resources that happens in order to safely fork, and as a result, can trigger a deadlock if new requests are made to the Runtime.

Add the ability to `drain` the engine's `Graph`, causing all work to fail fast with `Invalidated` the next time it attempts to interact with the `Graph`. Additionally, ensure that the `Invalidated` state is not actually stored, since it always represents an ephemeral failure.

(NB: We continue to push toward reducing usage of fork-without-exec by investing in the `--v2` UX, but removing it entirely might be a ways off.)

Entering the `fork_context` will cause all in-flight work to fail fast in a recoverable way, allowing for safe forking even in the presence of concurrent in-flight `pantsd` runs. According to ~30 minutes of runs that used to repro, this fixes #6426.