Add bootstrapper jar to compile the compile-bridge. (#6462)
### Problem

The `_zinc_cache_dir` option relies on the homedir existing to work, which is a blocker for remoting.
Discussion in #6155 .

### Solution

Implement a new Scala tool, `bootstrapper`, that takes the relevant jars as arguments and compiles the compiler bridge. This will be called, possibly remotely, from the `zinc` subsystem.

### Result

- A bit of refactoring of the `compiler` project.
- The creation of a `bootstrapper` target.

This PR is related to #6463

### Acknowledgements

This work was made together with @dotordogh and @illicitonion. Even if we used my fork for convenience, we programmed most of the code together. This work is based on @ity's work on the same problem.