skip integration test with pants_requirement() (#6493)
### Problem

See #6490. Currently, `test_pants_requirement_setup_requires_version` requires the version of pants specified in `src/python/pants/VERSION` to have been released to pypi. This makes it somewhat difficult for release candidates to pass CI.

### Solution

- Skip `test_pants_requirement_setup_requires_version` to unbreak release CI. This is the only references to the `pants_setup_requires` target I can find in the repo.

### Result

I don't necessarily think that depending on a `pants_requirement()` in a `python_dist()`'s `setup_requires` needs to be tested in CI, and I haven't figured out whether it's appropriate to delete this test entirely. Either way, that will be figured out in a followup diff.