Consume the bootstrapper and modify zinc to allow remote exec (#6463)
(This work was made together with @dotordogh and @illicitonion. Even if we used my fork for convenience, we programmed most of the code together. This work is based on @ity's work on the same problem.)

### Problem

See related issues below for discussion.

### Solution

We modify the `zinc` subsystem to call the newly created bootstraper jar.
That jar will compile the compiler-bridge that we pass to every invocation of zinc.
The compiled jar is cached in a fingerprinted directory, in the Pants working directory.
The compilation is made via an ExecuteProcessRequest, which will enable remote execution of the compilation.

### Result

Fixes #6155.