Fix encoding of workunits under pantsd (#6505)
### Problem

On python2 under `pantsd`, workunits containing unicode are currently not encoded to bytes because the synthetic stdout that is created in that environment does not have "std" in its name. 

### Solution

Ensure that all output files in use with `PlainTextReporter` accept writes of raw bytes, and then universally encode non-`binary_type` values. Mark an integration test `@ensure_daemon` to cover this.