[compile.rsc] Add strategy for compiling with Rsc and Zinc (#6408)
Rsc is an experimental Scala compiler focused on compilation speed (https://github.com/twitter/rsc).

Rsc implements so-called "outlining" for Scala, i.e. the process of computing type signatures for public, protected and private[package] members. These type signatures are saved to disk in SemanticDB format.

Rsc is also part of a toolchain that can produce jars containing type signatures in the format that the Scala compiler can understand. Using that tool chain we hope to be able to parallelize Scala compiles by first computing outlines for the dependency tree and then compiling the entire dependency tree in parallel.

This patch provides a first cut at implementing support for that toolchain.