Make the sizes of the members of `enum Node` more uniform (#6545)
### Problem

The clippy shard is currently failing with a notice that there is a large variance between the sizes of the members of the `Node` enum. It refers to https://rust-lang-nursery.github.io/rust-clippy/v0.0.212/index.html#large_enum_variant

According to `mem::size_of`, the sizes are currently:
```
  DigestFile(DigestFile),              // 32
  ExecuteProcess(ExecuteProcess),      // 200
  ReadLink(ReadLink),                  // 24
  Scandir(Scandir),                    // 24
  Select(Select),                      // 232
  Snapshot(Snapshot),                  // 16
  Task(Task),                          // 192
```

### Solution

Box the three largest members to appease clippy.

### Result

No appreciable difference in performance; happy clippy.