Switch synchronization primitive usage to parking_lot (#6564)
### Problem

As described in https://internals.rust-lang.org/t/standard-library-synchronization-primitives-and-undefined-behavior/8439 , the behaviour of rust's stdlib synchronization primitives can vary on different platforms, and we suspect that this could be making some synchronization issues harder to reproduce and track down.

### Solution
On a whim (and because `parking_lot` is very well respected), I decided to try porting pants to `parking_lot::{Mutex, RwLock}`, and received confirmation that a hang that some users were experiencing on Arch and Gentoo is resolved by this change.

### Result

Fixes #6510, and hopefully makes the behaviour of locking more consistent cross-platform. Performance remains roughly equivalent.