Change default for rebase map: part 1 (#6567)
Paired with @baroquebobcat on this 😄 

Scala part of the solution for #6434 

Now that the `jdk.home` is underneath the buildroot, we can change the rebase map to make it more general and to default to replacing the current working directory in the absolute path to this special `/proc/self/cwd`.

This allows analysis that is generated in a hermetic sandbox to be rebased correctly. Once the zinc wrapper is published, the python changes will come.