Move root executions off the main thread to allow for polling (#6570)
### Problem

The root futures of an engine request currently execute directly on the main thread using the built in "block this thread" executor. Having this work block the main thread makes it challenging to:

1. support doing other work with the main thread (which is the easiest place to render output), as in #6571.
2. support cleanly exiting when we are interrupted, as in #6368.

### Solution

Run root futures on the tokio runtime's executor, and have the futures send their result back to the main thread via an `mpsc` queue. The main thread can poll the queue at whatever frequency it wants while simultaneously checking for kill switches or rendering output.