[rsc-compile] update jdk dist lookup and usage to work remotely (#6593)
Remote execution assumes a .jdk that contains the jdk. Before, RscCompile would just use the local jdk paths, which would be wrong on a remote worker.

This patch introduces a small shim class to provide a distribution that creates remote compatible paths while still ensuring that the path lookup works correctly. It also changes the desandboxification helper to work for both local and remote exec.