recalculate dependencies in native link task
The indirect method of "caching" dependent target calculation we had previously resulted in
`LinkSharedLibraries` not incorporating the value of e.g. `--strict-deps` into its task
fingerprint. This fell apart because although we allow C and C++ targets to depend on each
other (and therefore `--strict-deps` being different for those target types is undefined -- this led
to errors).

We collapse the `NativeCompileSettings` subclasses (which take care of calculating dependencies given
the value of `--strict-deps`) into a single `NativeBuildSettings` subsystem, which is consumed by
compile and link tasks, and calculate the dependencies twice (which there was no reason we shouldn't
have been doing before).