Add support for un-cacheable rules, and stop caching console_rules (#6516)
### Problem

`@console_rule` tasks were cached, as were `@rule`s. This leads to undesired behaviour, such as empty output on non-first runs of `./pants list`.

### Solution

To include a field `cacheable` to console rules, that gets passed down to the engine and used to trigger the execution every time the rule is invoked. This entails:
- Modifying `_make_rule` in python to pass an extra parameter `cacheable`, with `True` by default and `False` for `@console_rule`s
- Augmenting `TaskRule` with the `cacheable` field.
- Augmenting `tasks_task_begin` in the FFI with an `int` field to represent `cacheable`. This int gets coerced to rust booleans.
- Creating a method `fn cacheable` in `Node` that uses the existing (unused) `cacheable` flag.
- Modifying `Entry::complete` to account for non-cacheable nodes.

### Result

Executing a `@console_rule` multiple times with pantsd enabled produces the same output.

Fixes #6146.