Ensure that changing platforms invalidates pex binary creation (#6202)
Currently, the platforms used to build pexes are not fingerprinted. This means that running ./pants binary some/python:bin twice with different platform settings will not recreate the pex on the second run. Instead it will retain the platform settings of the first run. This forces clean-alls when changing platform flags.

This patch causes the platforms flag to be fingerprinted and adds its owning subsystem to python binary create's subsystem dependencies.