Fix bugs in the parent/child relationship in ProcessManager (#6670)
### Problem

`ProcessManager.daemonize` and `ProcessManager.daemon_spawn` were out of sync with regard to the definition of what the `post_fork_parent` process was. `daemon_spawn` treated the original parent process of the double fork as the "post fork parent", while `daemonize` treated the short-lived "middle" process of the double fork as the "post fork parent".

Additionally, the middle process in a double-fork calls `exit` immediately after having forked its child, so having any sort of callback there was unlikely to be useful.

### Solution

Align `daemon_spawn` and `daemonize` with regard to the definition of the "post fork parent", such that both of those methods run the `post_fork_parent` callback in the original process.

### Result

The `post_fork_parent` method is called in the expected (and more useful) location.