Fixup `Checkstyle` local resolves.
Previously the local resolve in tests was pointed to the wrong path.
This was masked on unstable master since there, version changes come
earlier than code changes, meaning the checker dist from the prior
release was being remotely resolved. On a version bump change this
masking was exposed since no remote dist of the just-bumped-version was
deployed.

Upon unmasking, a bug was also revealed in the production code
implementation of local resolves from sys.path.

Both are fixed.

Fixes #6706