[fix] Fix `test_get_subprocess_output_interleaved` in python 3.
This one need some explaining.

This arised because Python 3 makes `sys.std{err,out}` a
`TextIOWrapper` and as per the python 3 man page "The text I/O layer
will still be line-buffered.". And that's with `-u` or
`PYTHONUNBUFFERED` set.

BUT, this test is not checking that python is buffering or not. It's
checking that if we run an executable in a subrocess for which we
capture output, we're correctly reading the potentially interleaved
STDIN and STDERR.

Unfortunatly, the "executable" we're using to test the capturing was
broken by the change to python 3. This change fixes the executable.