Restore valid imports
The original solution of a custom backport involved changing `import collections` to the style `from collections import x`. While this is not wrong, it was unnecessary in a couple files and led to a bigger diff than we want.

After this commit, the PR should be as simple as possible to achieve the original goal intended of fixing problematic uses of `future.moves.collections`.