Fix bad usage of `future.moves.collections` (#6747)
### Problem
`future.moves.collections` doesn't actually work as we intended. In Python 3, abstract classes like `Mapping` and `MutableSequence` were moved from `collections` to `collections.abc`. When we were using `future.moves.collections`, we expected to be able to import these values like `Iterable`.

However, the way the [source code](https://github.com/PythonCharmers/python-future/blob/master/src/future/moves/collections.py) is written for `future.moves.collections`, it doesn't grab any of the values from `collections.abc`, so we were getting `AttributeError`s when trying to use `future.moves.collections.Iterable`.  

### Solution
Ban use of `future.moves.collections` and only allow use of standard library's `collections`, which has aliases to `collections.abc` in Python 3. 

Note that this leads to a deprecation warning in Python 3.7, but we seem to not surface that deprecation. In Python 3.8, the aliases will cease to exist, which is okay because we should be done with Python 2 by then.