Make it easy to get a logger from a RunTracker instance. (#6771)
This allows anything with access to subsystems to get a logger,
with RunTracker.global_instance().logger.

Eventually we can leverage this to stop having to plumb loggers
in via the task context all over the place (or not logging
because we can't be bothered to do the plumbing). Logging is a
crosscutting concern, so it makes total sense for a logger to
be available in this manner.