Add prelude and epilogue (#6784)
### Problem

It would be nice to execute code before and after children are visited in graph traversals. The context of this change is the implementation of Dependency Banning (#6541), and the propagation of constraints through the dependency graph.

### Solution

Add optional prelude and epilogue function parameters to the dep. graph walk functions, that get executed right before visiting children. This should not incur any API or performance changes for existing code.

### Result

Calls like these are now possible:
```
a = self.make_target('a')
b = self.make_target('b', dependencies=[a])
names = []
self.build_graph.walk_transitive_dependency_graph(
    [b.address],
    work=lambda x: names.append(x.name),
    postorder=postorder,
    prelude=lambda x: names.append(x.name + "-pre"),
     epilogue=lambda x: names.append(x.name + "-epi"),
)
assertEqual(names, ['b', 'b-pre', 'a', 'a-pre', 'a-epi', 'b-epi'])
```

Blocks: #6541