Use both the deprecated and new locations of fatal_warnings args (#6798)
### Problem

Currently it is necessary to empty the `fatal_warnings_(enabled|disabled)_args` list in order for the values specified in `compiler_options_sets` to be used. But emptying the list (rather than using its default) requires actually specifying the deprecated options, which triggers a deprecation warning.

### Solution

Include either the values from `fatal_warnings_(enabled|disabled)_args` and `compiler_options_sets` depending on which has been explicitly set, to ensure that the options are consumed in either situation.

### Result

It's possible to migrate to `compiler_options_sets` without triggering a deprecation warning.