Make PythonInterpreterCache into a subsystem. (#6765)
It effectively depended on other subsystems (expecting instances to be passed into it), so it's a natural fit.

Switches it to use a local logger instead of one passed in from task context. The logger was only used for debug output anyway, and we want all such output to go through standard logging in the future anyway.

Also simplifies the interface of PythonInterpreterCache.setup(). Previously you could pass a list of paths to it, overriding its internal logic. However this was only used in tests, which can easily be converted to use the appropriate option instead.

This is the first step in simplifying the path-selection logic to be entirely option-based, instead of the current chain of ors that can't be fully overridden in options.