Fix a bug when selecting interpreters with no constraints at all. (#6822)
Previously, if there were no interpreter constraints at all, not even default ones (e.g., if the value of `interpreter_constraints` in `[python-setup]` was explicitly set to an empty list in pants.ini) we would fail to select an interpreter. This was due to the empty set of interpreters vacuously satisfying the condition of "not leaving any filters unsatisfied" when there are no filters.

This change fixes this, and updates the test that was supposed to exercise this to actually do so. Previously that test set the constraint list to `['']`, but that is a contrived value, and not the same as having no values at all in the list.