Port remaining tests to new TestBase base class. (#6864)
This ports all remaining direct users of the old BaseTest to the new base class. It does not deal with remaining uses of the old pants_test.tasks.TaskTestBase, which extends BaseTest. Only after we do that can we remove BaseTest.

This change also gets rid of BuildFileTestBase. It was only used by one test, FilesystemBuildFileTest, so I inlined the necessary code there. Note that I'm not even sure we still need FilesystemBuildFileTest.