Expose an API to pass multiple Params to an engine request (#6871)
### Problem

Although consumption of multiple `Params` is now supported, there is currently no way to actually provide them to a request. See #6478 for more info.

### Solution

Expose and test the ability to pass in a small `Params` datatype as a subject, which is flattened into a corresponding set of rust-side `Params` in `add_root_select`. 

### Result

Fixes #6478.